IDPKit -> removed the warnings OSX -> Removed NSBundle category from osx IDPCoreDataManager -> made portable

+++ Module/FoundationKit/Shared/Categories/NSArray/NSArray+IDPExtensions.m
@@ -42,7 +42,7 @@
 }
 
 - (NSUInteger)randomWithCount:(NSUInteger)count {
-    return arc4random_uniform(count);
+    return arc4random_uniform((uint32_t)count);
 }
 
 @end
